wip: extract documentation ext point
