+++ src/vs/workbench/contrib/documentation/common/documentationExtensionPoint.ts
