+++ src/vs/workbench/workbench.common.main.ts
@@ -272,6 +272,9 @@ import 'vs/workbench/contrib/userDataSync/browser/userDataSync.contribution';
 // Code Actions
 import 'vs/workbench/contrib/codeActions/common/codeActions.contribution';
 
+// Documentation
+import 'vs/workbench/contrib/documentation/common/documentation.contribution';
+
 // Timeline
 import 'vs/workbench/contrib/timeline/browser/timeline.contribution';
